move distribute_text func from make_sen-.py to para-.py
