+++ source/make_sentence_block.py
@@ -193,42 +193,3 @@ def make_sentence_block(sentence_data: dict, threshold: float = 1.5) -> dict:
     flush_block()
     
     return result
-
-
-def distribute_text(text: str, line_num: int, line_width: list) -> list:
-    """
-    Distribute a text string into multiple lines according to specified line widths.
-
-    The text is split into words, and the number of words allocated to each line is determined
-    by the proportional width of that line. The final line receives any remaining words.
-
-    Args:
-        text (str): The text to distribute.
-        line_num (int): The number of lines to divide the text into.
-        line_width (list): A list of integers representing the widths for each line.
-
-    Returns:
-        list: A list of strings, each representing a line of text.
-    """
-    tokens = text.split()
-    total_tokens_length = len(tokens)
-    total_width = sum(line_width)
-
-    # Calculate the proportion of tokens for each line.
-    line_ratios = [round(w / total_width, 2) for w in line_width]
-
-    distributed_lines = []
-
-    # Divide one sentence into multiple lines to match the line width ratio
-    for i in range(line_num):
-        if i == line_num - 1:
-            line_tokens = tokens
-            tokens = []
-        else:
-            num_tokens = int(total_tokens_length * line_ratios[i])
-            line_tokens = tokens[:num_tokens]
-            tokens = tokens[num_tokens:]
-
-        distributed_lines.append(" ".join(line_tokens))
-    
-    return distributed_lines
